+++ commands/invoke.go
@@ -70,7 +70,7 @@ func runInvoke(cmd *cobra.Command, args []string) error {
 		}
 	}
 
-	gatewayAddress := getGatewayURL(gateway, defaultGateway, yamlGateway)
+	gatewayAddress := getGatewayURL(gateway, defaultGateway, yamlGateway, os.Getenv("OPENFAAS_URL"))
 
 	stat, _ := os.Stdin.Stat()
 	if (stat.Mode() & os.ModeCharDevice) != 0 {
