+++ sample/hmac.yml
@@ -1,6 +1,6 @@
 provider:
   name: faas
-  gateway: http://localhost:8080
+  gateway: http://127.0.0.1:8080
 
 functions:
   hmac:
