+++ stack/stack_test.go
@@ -12,7 +12,7 @@ import (
 
 const TestData_1 string = `provider:
   name: faas
-  gateway: http://localhost:8080
+  gateway: http://127.0.0.1:8080
   network: "func_functions"
 
 functions:
@@ -45,7 +45,7 @@ functions:
 
 const TestData_2 string = `provider:
   name: faas
-  gateway: http://localhost:8080
+  gateway: http://127.0.0.1:8080
   network: "func_functions"
 
 `
